jarvis-ear/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis-ear"
version = "0.1.0"
description = "Always-on audio capture daemon for Jarvis voice assistant"
requires-python = ">=3.11"
dependencies = [
    "pyalsaaudio>=0.11",
    "onnxruntime>=1.19",
    "openwakeword>=0.4",
    "silero-vad>=5.1",
    "python-socketio[client]>=5.16",
]

[project.scripts]
jarvis-ear = "jarvis_ear.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
